+++ src/app/app.component.ts
@@ -1,5 +1,7 @@
 import { Component } from '@angular/core';
 import { KeycloakService } from 'keycloak-angular';
+import { Observable } from 'rxjs';
+import { LoginStatusService } from './services/login-status.service';
 
 @Component({
   selector: 'app-root',
@@ -7,10 +9,16 @@ import { KeycloakService } from 'keycloak-angular';
   styleUrls: ['./app.component.scss'],
 })
 export class AppComponent {
-  constructor(private keycloak: KeycloakService) {}
+  constructor(
+    private keycloak: KeycloakService,
+    loginStatusService: LoginStatusService
+  ) {
+    this.isLoggedIn = loginStatusService.getIsLoggedIn();
+  }
   title = 'user-web-client';
+  isLoggedIn: Observable<boolean>;
 
   async onLogout() {
-    this.keycloak.logout();
+    await this.keycloak.logout();
   }
 }