+++ src/app/services/login-status.service.ts
@@ -1,0 +1,20 @@
+import { Injectable } from '@angular/core';
+import { BehaviorSubject, Observable } from 'rxjs';
+import { distinctUntilChanged } from 'rxjs/operators';
+
+@Injectable({
+  providedIn: 'root',
+})
+export class LoginStatusService {
+  constructor() {}
+
+  private isLoggedIn$ = new BehaviorSubject<boolean>(false);
+
+  setLoggedIn(isLoggedIn: boolean) {
+    this.isLoggedIn$.next(isLoggedIn);
+  }
+
+  getIsLoggedIn(): Observable<boolean> {
+    return this.isLoggedIn$.pipe(distinctUntilChanged());
+  }
+}